Removing superflouos styles and javascript
CTEFE doesn't require any of the code and files being removed.
I have left /assets/javascripts/show-hide-content.js in. Although we're not using anything from here, I think we maybe should because there's potentially some useful accessibility code in there that we 'should' be using. I will need to investigate further if this is useful or there is a better and established (elsewhere) pattern for making our hidden content and progress disclosures more accessible with aria.

- I have left in form.scss but removed everything styles for error messaging
– removed all style imports of files that I have deleted
- removed all unnecessary styles
- removed all js from /assets/javascripts/companyregistrationeligibilityfrontend.js