Update service and add styles to adjust header layout
This work updates the service name to 'Set up a limited company and register for Corporation Tax'. This however extends the service name further across the page, thus forcing the 'sign in/out' link to drop below, breaking the layout.

I have added some styles which adjust the layout from a 1/3 | 2/3's split to a 20% | 80% split and set the logo with a minimum width so that doesn't break in the lower screen width range of the tablet breakpoint.